Add color mode switcher
